feat: refactor data fetching; add cron to visit pages
