+++ vercel.json
@@ -1,6 +1,6 @@
 {
   "crons": [{
     "path": "/api/cron",
-    "schedule": "0 10 * * *"
+    "schedule": "30 * * * *"
   }]
 }
